Attempt to get rid of compiler warnings

+++ fs2_open/code/missionui/missionweaponchoice.cpp
@@ -9,13 +9,16 @@
 
 /*
  * $Logfile: /Freespace2/code/MissionUI/MissionWeaponChoice.cpp $
- * $Revision: 2.35 $
- * $Date: 2005-02-13 08:41:25 $
+ * $Revision: 2.36 $
+ * $Date: 2005-02-18 05:24:26 $
  * $Author: wmcoolmon $
  *
  * C module for the weapon loadout screen
  *
  * $Log: not supported by cvs2svn $
+ * Revision 2.35  2005/02/13 08:41:25  wmcoolmon
+ * 3D models in weapons selection screen and nonstandard resolution fixes for ship selection screen.
+ *
  * Revision 2.34  2005/02/08 23:49:59  taylor
  * update/add .cvsignore files for project file changes
  * silence warning about depreciated strings.h stuff for MSVC 2005
@@ -3384,9 +3387,9 @@ void weapon_select_do(float frametime)
 		//Find the center of teh submodel
 		weap_closeup.xyz.x = -(bs->min.xyz.z + (bs->max.xyz.z - bs->min.xyz.z)/2.0f);
 		weap_closeup.xyz.y = bs->min.xyz.y + (bs->max.xyz.y - bs->min.xyz.y)/2.0f;
-		weap_closeup.xyz.z = (weap_closeup.xyz.x/tan(zoom / 2.0f));
+		weap_closeup.xyz.z = (weap_closeup.xyz.x/tan((float)(zoom / 2.0f)));
 
-		y_closeup = -(weap_closeup.xyz.y/tan(zoom / 2.0f));
+		y_closeup = -(weap_closeup.xyz.y/tan((float)(zoom / 2.0f)));
 		if(y_closeup < weap_closeup.xyz.z)
 		{
 			weap_closeup.xyz.z = y_closeup;
